Move Long Polling and SSE send logic to SendUtils
