+++ src/Microsoft.AspNetCore.Sockets.Client/ServerSentEventsTransport.cs
@@ -55,7 +55,7 @@ namespace Microsoft.AspNetCore.Sockets.Client
             _application = application;
             var sseUrl = Utils.AppendPath(url, "sse");
             var sendUrl = Utils.AppendPath(url, "send");
-            var sendTask = SendMessages(sendUrl, _transportCts.Token);
+            var sendTask = SendUtils.SendMessages(sendUrl, _application, _httpClient, _transportCts, _logger);
             var receiveTask = OpenConnection(_application, sseUrl, _transportCts.Token);
 
             Running = Task.WhenAll(sendTask, receiveTask).ContinueWith(t =>
@@ -125,101 +125,6 @@ namespace Microsoft.AspNetCore.Sockets.Client
             finally
             {
                 _transportCts.Cancel();
-            }
-        }
-
-        private async Task SendMessages(Uri sendUrl, CancellationToken cancellationToken)
-        {
-            _logger.LogInformation("Starting the send loop");
-
-            List<SendMessage> messages = null;
-            try
-            {
-                while (await _application.Input.WaitToReadAsync(cancellationToken))
-                {
-                    messages = new List<SendMessage>();
-                    while (!cancellationToken.IsCancellationRequested && _application.Input.TryRead(out SendMessage message))
-                    {
-                        messages.Add(message);
-                    }
-
-                    if (messages.Count > 0)
-                    {
-                        _logger.LogDebug("Sending {messageCount} message(s) to the server using url: {url}", messages.Count, sendUrl);
-
-                        var request = new HttpRequestMessage(HttpMethod.Post, sendUrl);
-                        request.Headers.UserAgent.Add(DefaultUserAgentHeader);
-
-                        var memoryStream = new MemoryStream();
-
-                        var pipe = memoryStream.AsPipelineWriter();
-                        var output = new PipelineTextOutput(pipe, TextEncoder.Utf8);
-                        await WriteMessagesAsync(messages, output, MessageFormat.Binary);
-
-                        memoryStream.Seek(0, SeekOrigin.Begin);
-
-                        request.Content = new StreamContent(memoryStream);
-                        request.Content.Headers.ContentType = MediaTypeHeaderValue.Parse(MessageFormatter.GetContentType(MessageFormat.Binary));
-
-                        var response = await _httpClient.SendAsync(request);
-                        response.EnsureSuccessStatusCode();
-
-                        _logger.LogDebug("Message(s) sent successfully");
-                        foreach (var message in messages)
-                        {
-                            message.SendResult?.TrySetResult(null);
-                        }
-                    }
-                }
-            }
-            catch (OperationCanceledException)
-            {
-                _logger.LogError("Send cancelled");
-
-                if (messages != null)
-                {
-                    foreach (var message in messages)
-                    {
-                        message.SendResult?.TrySetCanceled();
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogDebug("Error while sending to '{url}' : '{exception}'", sendUrl, ex);
-                if (messages != null)
-                {
-                    foreach (var message in messages)
-                    {
-                        message.SendResult?.TrySetException(ex);
-                    }
-                }
-                throw;
-            }
-            finally
-            {
-                // Make sure the poll loop is terminated
-                _transportCts.Cancel();
-            }
-
-            _logger.LogInformation("Send loop stopped");
-        }
-
-        private async Task WriteMessagesAsync(List<SendMessage> messages, PipelineTextOutput output, MessageFormat format)
-        {
-            output.Append(MessageFormatter.GetFormatIndicator(format), TextEncoder.Utf8);
-
-            foreach (var message in messages)
-            {
-                _logger.LogDebug("Writing '{messageType}' message to the server", message.Type);
-
-                var payload = message.Payload ?? Array.Empty<byte>();
-                if (!MessageFormatter.TryWriteMessage(new Message(payload, message.Type, endOfMessage: true), output, format))
-                {
-                    // We didn't get any more memory!
-                    throw new InvalidOperationException("Unable to write message to pipeline");
-                }
-                await output.FlushAsync();
             }
         }
